Renamed 'to_gff' to 'parser_factory' and updated docstrings. Restored 'pacbio.bam' in the sample_data, fixing the system tests
