+++ Mikado/preparation/annotation_parser.py
@@ -1,5 +1,5 @@
 import multiprocessing
-from ..parsers import to_gff
+from ..parsers import parser_factory
 from ..parsers.bam_parser import BamParser
 from ..utilities.log_utils import create_queue_logger
 from ..utilities import overlap
@@ -739,7 +739,7 @@ class AnnotationParser(multiprocessing.Process):
                               strand_specific,
                               shelf_name)
             try:
-                gff_handle = to_gff(handle)
+                gff_handle = parser_factory(handle)
                 loader = loaders.get(gff_handle.__annot_type__, None)
                 if loader is None:
                     raise ValueError("Invalid file type: {}".format(gff_handle.name))
